+++ CHANGES.rst
@@ -4,7 +4,7 @@ N&S Dependency Loader changelog
 1.0.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Bumped threedigrid-builder to 1.24.1
 
 
 1.0.4 (2025-04-11)